SWY_server/requirements.txt
fastapi
uvicorn
motor
rembg
Pillow
google-generativeai
bcrypt
python-multipart
